+++ conftest.py
@@ -1,5 +1,3 @@
-import os
-
 import pytest
 from sphinx.testing.path import path
